used server side fetchign as in the code review
