+++ src/components/pages/home/index.tsx
@@ -14,6 +14,7 @@ const Home = async () => {
     latestBlocks,
     latestTransactions,
     latestTransactionsEnqueued,
+    transactionHistory,
   } = await fetchHomePageData();
   return (
     <main className="flex flex-1 flex-col gap-4 p-4 md:gap-4 md:p-4">
@@ -26,7 +27,7 @@ const Home = async () => {
           <LatestBlockAndTxs blockNumber={latestBlocks[0]?.number} />
           <LatestL1TxBatch />
         </div>
-        <TransactionHistory />
+        <TransactionHistory history={transactionHistory} />
       </div>
       <div className="grid gap-4 md:gap-4 lg:grid-cols-2 xl:grid-cols-3">
         <LatestBlocks blocks={latestBlocks} />
