+++ src/lib/fetch-data.ts
@@ -1,5 +1,5 @@
 import { range } from "lodash";
-import { subDays, formatISO } from "date-fns";
+import { subDays, formatISO, format } from "date-fns";
 import { Hash } from "viem";
 import {
   extractTransactionDepositedLogs,
@@ -20,6 +20,33 @@ import portal from "@/lib/contracts/portal/contract";
 import l1CrossDomainMessenger from "@/lib/contracts/l1-cross-domain-messenger/contract";
 import { prisma } from "@/lib/prisma";
 import { getSignatureBySelector } from "@/lib/4byte-directory";
+
+export const fetchTransactionHistory = async () => {
+  const dates = Array.from({ length: 15 }, (_, i) => {
+    const date = subDays(new Date(), 14 - i);
+    return {
+      name: format(date, "MMM d"),
+      formattedDate: format(date, "yyyy-MM-dd"),
+    };
+  });
+
+  const responses = await Promise.all(
+    dates.map(async (dateObj) => {
+      const response = await fetch(
+        `https://api.coinbase.com/v2/prices/OP-USD/spot?date=${dateObj.formattedDate}`,
+      );
+      const json = await response.json();
+      return {
+        name: dateObj.name,
+        date: dateObj.formattedDate,
+        price: parseFloat(json.data.amount),
+        transactionsCount: Math.floor(Math.random() * 200) + 400,
+      };
+    }),
+  );
+
+  return responses;
+};
 
 export const fetchLatestBlocks = async (start: bigint): Promise<Block[]> => {
   const blocksPerPage = BigInt(process.env.NEXT_PUBLIC_BLOCKS_PER_PAGE);
@@ -217,6 +244,7 @@ export const fetchHomePageDataFromJsonRpc = async () => {
     latestBlocks,
     { transactions: latestTransactions },
     { transactions: latestTransactionsEnqueued },
+    transactionHistory,
   ] = await Promise.all([
     fetchTokensPrices(),
     fetchLatestBlocks(latestL2BlockNumber),
@@ -226,12 +254,14 @@ export const fetchHomePageDataFromJsonRpc = async () => {
       "0x",
       latestL1BlockNumber,
     ),
+    fetchTransactionHistory(),
   ]);
   return {
     tokensPrices,
     latestBlocks: latestBlocks.slice(0, 6),
     latestTransactions: latestTransactions.slice(0, 6),
     latestTransactionsEnqueued: latestTransactionsEnqueued.slice(0, 10),
+    transactionHistory,
   };
 };
 
@@ -241,6 +271,7 @@ export const fetchHomePageDataFromDatabase = async () => {
     latestBlocks,
     latestTransactions,
     latestTransactionsEnqueued,
+    transactionHistory,
   ] = await Promise.all([
     fetchTokensPrices(),
     prisma.block.findMany({
@@ -256,6 +287,7 @@ export const fetchHomePageDataFromDatabase = async () => {
       orderBy: { timestamp: "desc" },
       take: 10,
     }),
+    fetchTransactionHistory(),
   ]);
   return {
     tokensPrices,
@@ -264,6 +296,7 @@ export const fetchHomePageDataFromDatabase = async () => {
     latestTransactionsEnqueued: latestTransactionsEnqueued.map(
       fromPrismaTransactionEnqueued,
     ),
+    transactionHistory,
   };
 };
